Update 15:00/4/2/2020
Fixed a typo in Spot3D ("locaion" to "location")

Finished an axis-angle constructor for the Quaternion class, which uses the new Spot3D class.
Made the proper version of toEulerAnglesXYZ the actual version used (formerly, it has "2" at the end of the name while the other remained)

Created QuaternionTester5, which tests the new Axis-Angle constructor.