updater 15:47/5/7/2020
Updated SlerpModelRotate2 and SlerpModelRotate2_2 to get rid of remaining fragments and name usage from the Computer Graphics project and SlerpModelRotate1 that I derived it from.
Made SlerpModelRotate2_2, which handles updating positions differently so dragging during rotation only updates the "current" position, leaving the other alone so the model can actually catch up.
Updated the Quaternion fromAxisAngle function to return an identity Quaternion if the angle passed in is zero, as this otherwise was an edge case that created a quaterion that would break the model rotation program.